⭐ New Features 1. 适配 Spring boot 4.0.0-M3，删除 org.hibernate.annotations.Comment 使用 jakarta.persistence 自带方法

+++ src/main/java/com/simonvonxcvii/turing/model/dto/PermissionDTO.java
@@ -36,7 +36,7 @@ public class PermissionDTO extends PageQuery {
      * 排序编号
      */
     @NotNull(message = "排序编号不能为空")
-    private Integer sort;
+    private Short sort;
     /**
      * 子级权限集合
      */
